chore: distribute cache functions
