+++ apps/server/src/services/rundown-service/cachedRundown.utils.ts
@@ -11,9 +11,11 @@ import {
 } from 'ontime-types';
 import { swapOntimeEvents } from 'ontime-utils';
 
+import { calculateRuntimeDelays, calculateRuntimeDelaysFromIndex, getDelayAt } from './rundown.utils.js';
+
+import { isProduction } from '../../setup.js';
 import { DataProvider } from '../../classes/data-provider/DataProvider.js';
 import { getCached, runtimeCacheStore } from '../../stores/cachingStore.js';
-import { isProduction } from '../../setup.js';
 import { deleteAtIndex, insertAtIndex, reorderArray } from '../../utils/arrayUtils.js';
 import { _applyDelay } from '../delayUtils.js';
 
@@ -154,12 +156,6 @@ export async function cachedEdit(
   return newEvent;
 }
 
-export async function cachedBatchEdit(ids: string[], patchObject: Partial<OntimeEvent>) {
-  const cachedEdits = ids.map((id) => cachedEdit(id, patchObject));
-
-  await Promise.allSettled(cachedEdits);
-}
-
 /**
  * Deletes an event with given id from rundown, ensuring replication to delayed rundown cache
  * @param eventId
@@ -229,6 +225,12 @@ export async function cachedReorder(eventId: string, from: number, to: number) {
   return reorderedEvent;
 }
 
+
+export async function cachedBatchEdit(ids: string[], patchObject: Partial<OntimeEvent>) {
+  const cachedEdits = ids.map((id) => cachedEdit(id, patchObject));
+
+  await Promise.allSettled(cachedEdits);
+}
 export async function cachedClear() {
   await DataProvider.clearRundown();
   runtimeCacheStore.setCached(delayedRundownCacheKey, []);
@@ -277,93 +279,4 @@ export async function cachedApplyDelay(eventId: string) {
   await DataProvider.setRundown(persistedRundown);
 
   rundownRevision++;
-}
-
-/**
- * Calculates all delays in a given rundown
- * @param rundown
- */
-export function calculateRuntimeDelays(rundown: OntimeRundown) {
-  let accumulatedDelay = 0;
-  const updatedRundown = [...rundown];
-
-  for (const [index, event] of updatedRundown.entries()) {
-    if (isOntimeDelay(event)) {
-      accumulatedDelay += event.duration;
-    } else if (isOntimeBlock(event)) {
-      accumulatedDelay = 0;
-    } else if (isOntimeEvent(event)) {
-      updatedRundown[index] = {
-        ...event,
-        delay: accumulatedDelay,
-      };
-    }
-  }
-  return updatedRundown;
-}
-
-/**
- * Calculate delays in rundown from a given index
- * @param eventIndex
- * @param rundown
- */
-export function calculateRuntimeDelaysFromIndex(eventIndex: number, rundown: OntimeRundown) {
-  if (eventIndex === -1) {
-    throw new Error('ID not found at index');
-  }
-
-  let accumulatedDelay = getDelayAt(eventIndex, rundown);
-  const updatedRundown = [...rundown];
-
-  for (let i = eventIndex; i < rundown.length; i++) {
-    const event = rundown[i];
-    if (isOntimeDelay(event)) {
-      accumulatedDelay += event.duration;
-    } else if (isOntimeBlock(event)) {
-      if (i === eventIndex) {
-        accumulatedDelay = 0;
-      } else {
-        break;
-      }
-    } else if (isOntimeEvent(event)) {
-      updatedRundown[i] = {
-        ...event,
-        delay: accumulatedDelay,
-      };
-    }
-  }
-  return updatedRundown;
-}
-
-/**
- * Calculate delays in rundown from an event with given id
- * @param eventId
- * @param rundown
- */
-export function calculateRuntimeDelaysFrom(eventId: string, rundown: OntimeRundown) {
-  const index = rundown.findIndex((event) => event.id === eventId);
-  return calculateRuntimeDelaysFromIndex(index, rundown);
-}
-
-/**
- * Calculates delay to an event at a given index
- * @param eventIndex
- * @param rundown
- */
-export function getDelayAt(eventIndex: number, rundown: OntimeRundown): number {
-  if (eventIndex < 1) {
-    return 0;
-  }
-
-  // we need to check the event before
-  const event = rundown[eventIndex - 1];
-
-  if (isOntimeDelay(event)) {
-    return event.duration + getDelayAt(eventIndex - 1, rundown);
-  } else if (isOntimeBlock(event)) {
-    return 0;
-  } else if (isOntimeEvent(event)) {
-    return event.delay ?? 0;
-  }
-  return 0;
 }
